requirements.txt
playwright==1.47.0
pandas==2.2.2
python-dateutil==2.9.0
